[Silver III] Title: 주유소, Time: 328 ms, Memory: 41428 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -3,14 +3,14 @@ import java.util.*;
 
 class Main {
     static int N;
-    static int[] price, distance;
+    static long[] price, distance;
     public static void main(String[] args) throws Exception {
         set();
         pro();
     }
     public static void pro(){
-        int nowPrice = price[0];
-        int sum = nowPrice * distance[0];
+        long nowPrice = price[0];
+        long sum = nowPrice * distance[0];
         for (int i = 1; i < N-1; i++) {
             nowPrice = Math.min(nowPrice, price[i]);
             sum += (nowPrice*distance[i]);
@@ -22,8 +22,8 @@ class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         N = Integer.parseInt(br.readLine());
         StringTokenizer st = new StringTokenizer(br.readLine());
-        distance = new int[N-1];
-        price = new int[N];
+        distance = new long[N-1];
+        price = new long[N];
         for (int i = 0; i < N-1; i++) {
             distance[i] = Integer.parseInt(st.nextToken());
         }
